+++ CHANGELOG.md
@@ -16,7 +16,9 @@
     Click to see more.
   </summary>
 
-
+### :house: Internal
+- `ricos-content`
+  - [#2610](https://github.com/wix/ricos/pull/2610) tiptap converters fixed
 
 </details>
 <hr/>